refactor: rename read_files_table -> read_path
