+++ mjolnir/input/read_system.hpp
@@ -6,7 +6,7 @@
 #include <mjolnir/util/throw_exception.hpp>
 #include <mjolnir/util/logger.hpp>
 #include <mjolnir/math/vector_util.hpp>
-#include <mjolnir/input/read_files_table.hpp>
+#include <mjolnir/input/read_path.hpp>
 
 namespace mjolnir
 {
